Rename utility files to `site-title` for clarity
